% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvRef.R
\name{cell_type_markers}
\alias{cell_type_markers}
\title{Generate markers from DEGs}
\usage{
cell_type_markers(
  path = "/your_path/",
  analysis_name = "test",
  top_deg_num = 50
)
}
\arguments{
\item{path}{A character of a single path to save all the results}

\item{analysis_name}{A character vector representing name of analyses}

\item{top_deg_num}{A numeric specify how many top DEGs will be selected}
}
\description{
This function will apply recursive feature elimination (RFE) with a random forest model to find representative marker genes.
It would return the marker genes for each cluster.
Txt files of the intermediate results can be downloaded in this sub-directory.
}
\examples{
cell_type_markers(path="/your_path/",analysis_name="test", 
top_deg_num=50)
}
