% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvRef.R
\name{gen_crm}
\alias{gen_crm}
\title{Generate DEGs}
\usage{
gen_crm(path = "/your_path/", analysis_name = "test")
}
\arguments{
\item{path}{A character of a single path to save all the results}

\item{analysis_name}{A character vector representing name of analyses}
}
\description{
This function will use the selected markers to build CRM
It would return the final CRM.
Txt files of the CRM can be downloaded in this sub-directory.
}
\examples{
cell_type_markers(path="/your_path/", analysis_name="test")
}
