% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvRef.R
\name{pseudobulk_deg}
\alias{pseudobulk_deg}
\title{Generate DEGs}
\usage{
pseudobulk_deg(
  analysis_name = "test",
  file = "data/test.nsclc.seur.meta.norm.qc.rds",
  path = "/your_path/",
  celltype = "cluster_annot",
  sample_id = "sample_name",
  top_deg_num = 50
)
}
\arguments{
\item{analysis_name}{A character vector representing name of analyses}

\item{file}{A character of a single path to scRNA-seq in rds format}

\item{path}{A character of a single path to save all the results}

\item{celltype}{A character specify the layer to the scRNA-seq to read cell type information}

\item{sample_id}{A character specify the layer to the scRNA-seq to read sample name information}

\item{top_deg_num}{A numeric specify how many top DEGs will be selected}
}
\description{
This function allows to upload their own scRNA-seq, build the pseudobulk and find DEGs.
It would return the pseudobulk files and DEGs from each cluster.
Txt files Txt files of the intermediate results can be downloaded in the sub-directory.
}
